code for breaking symmetry

+++ src/main/java/ga/operations/fitnessFunctions/GRNFitnessFunctionMultipleTargets.java
@@ -110,7 +110,7 @@ public class GRNFitnessFunctionMultipleTargets extends GRNFitnessFunction<Simple
             } while (currentRound < this.maxCycle && isNotStable);
 
             if (currentRound < maxCycle) {
-                int hammingDistance = this.getHammingDistance(currentAttractor, target);
+                double hammingDistance = this.getHammingDistance(currentAttractor, target);
                 double thisFitness = Math.pow((1 - (hammingDistance / ((double) target.length))), 5);
                 fitnessValue += thisFitness;
             } else {
